% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{ppn_charter_to_school}
\alias{ppn_charter_to_school}
\title{PPN charter to school}
\usage{
ppn_charter_to_school(charter, grade)
}
\arguments{
\item{charter}{one of: GPBX, GPLES, BPBX}

\item{grade}{one of: PK, K, 1, 2, 3, 4, 5, 6, 7, 8}
}
\value{
vector of school short codes
}
\description{
the state only understands our data as one school, but we have
an ES campus and an MS campus.  turn the generic charter name into the
school short code
}
