% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{ppn_names_to_codes}
\alias{ppn_names_to_codes}
\title{PPN names to codes}
\usage{
ppn_names_to_codes(name_vector)
}
\arguments{
\item{name_vector}{character vector of school names}
}
\value{
character vector of codes
}
\description{
PPN names to codes
}
\examples{
ex_sch <- c("Girls Prep Lower East Side Elementary School",
"Girls Prep Bronx Elementary School", "Boys Prep Elementary School")
ppn_names_to_codes(ex_sch)
}
